coprocessor/dag/expr: add some un-pushed builtin UDFs(Regexp) (#3196)
