+++ src/coprocessor/dag/expr/builtin_compare.rs
@@ -14,14 +14,11 @@
 use std::borrow::Cow;
 use std::cmp::Ordering;
 use std::i64;
-use std::slice::Iter;
 
 use super::{Error, EvalContext, Result, ScalarFunc};
 use coprocessor::codec::mysql::{Decimal, Duration, Json, Time};
 use coprocessor::codec::{datum, mysql, Datum};
 use coprocessor::dag::expr::Expression;
-
-const MAX_RECURSE_LEVEL: usize = 1024;
 
 #[derive(Clone, Copy, PartialEq)]
 pub enum CmpOp {
@@ -205,17 +202,6 @@ impl ScalarFunc {
     pub fn in_json(&self, ctx: &mut EvalContext, row: &[Datum]) -> Result<Option<i64>> {
         do_in(self, |v| v.eval_json(ctx, row), |l, r| Ok(l.cmp(r)))
     }
-
-    /// NOTE: LIKE compare target with pattern as bytes, even if they have different
-    /// charsets. This behaviour is for keeping compatible with TiDB. But MySQL
-    /// compare them as bytes only if any charset of pattern or target is binary,
-    /// otherwise MySQL will compare decoded string.
-    pub fn like(&self, ctx: &mut EvalContext, row: &[Datum]) -> Result<Option<i64>> {
-        let target = try_opt!(self.children[0].eval_string(ctx, row));
-        let pattern = try_opt!(self.children[1].eval_string(ctx, row));
-        let escape = try_opt!(self.children[2].eval_int(ctx, row)) as u32;
-        Ok(Some(like(&target, &pattern, escape, 0)? as i64))
-    }
 }
 
 fn do_compare<T, E, F>(mut e: E, op: CmpOp, get_order: F) -> Result<Option<i64>>
@@ -311,81 +297,12 @@ where
     ret_when_not_matched
 }
 
-// Do match until '%' is found.
-#[inline]
-fn partial_like(tcs: &mut Iter<u8>, pcs: &mut Iter<u8>, escape: u32) -> Option<bool> {
-    loop {
-        match pcs.next().cloned() {
-            None => return Some(tcs.next().is_none()),
-            Some(b'%') => return None,
-            Some(c) => {
-                let (npc, escape) = if u32::from(c) == escape {
-                    pcs.next().map_or((c, false), |&c| (c, true))
-                } else {
-                    (c, false)
-                };
-                let nsc = match tcs.next() {
-                    None => return Some(false),
-                    Some(&c) => c,
-                };
-                if nsc != npc && (npc != b'_' || escape) {
-                    return Some(false);
-                }
-            }
-        }
-    }
-}
-
-fn like(target: &[u8], pattern: &[u8], escape: u32, recurse_level: usize) -> Result<bool> {
-    let mut tcs = target.iter();
-    let mut pcs = pattern.iter();
-    loop {
-        if let Some(res) = partial_like(&mut tcs, &mut pcs, escape) {
-            return Ok(res);
-        }
-        let next_char = loop {
-            match pcs.next().cloned() {
-                Some(b'%') => {}
-                Some(b'_') => if tcs.next().is_none() {
-                    return Ok(false);
-                },
-                // So the pattern should be some thing like 'xxx%'
-                None => return Ok(true),
-                Some(c) => {
-                    break if u32::from(c) == escape {
-                        pcs.next().map_or(escape, |&c| u32::from(c))
-                    } else {
-                        u32::from(c)
-                    };
-                }
-            }
-        };
-        if recurse_level >= MAX_RECURSE_LEVEL {
-            // TODO: maybe we should test if stack is actually about to overflow.
-            return Err(box_err!(
-                "recurse level should not be larger than {}",
-                MAX_RECURSE_LEVEL
-            ));
-        }
-        // Pattern must be something like "%xxx".
-        loop {
-            let s = match tcs.next() {
-                None => return Ok(false),
-                Some(&s) => u32::from(s),
-            };
-            if s == next_char && like(tcs.as_slice(), pcs.as_slice(), escape, recurse_level + 1)? {
-                return Ok(true);
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
     use coprocessor::codec::mysql::{Decimal, Duration, Json, Time};
     use coprocessor::codec::Datum;
-    use coprocessor::dag::expr::test::{col_expr, datum_expr, scalar_func_expr};
+    use coprocessor::dag::expr::test::col_expr;
     use coprocessor::dag::expr::{EvalContext, Expression};
     use protobuf::RepeatedField;
     use std::{i64, u64};
@@ -625,53 +542,6 @@ mod test {
             let e = Expression::build(&mut ctx, expr).unwrap();
             let res = e.eval(&mut ctx, &row).unwrap();
             assert_eq!(res, exp);
-        }
-    }
-
-    #[test]
-    fn test_like() {
-        let cases = vec![
-            (r#"hello"#, r#"%HELLO%"#, '\\', false),
-            (r#"Hello, World"#, r#"Hello, World"#, '\\', true),
-            (r#"Hello, World"#, r#"Hello, %"#, '\\', true),
-            (r#"Hello, World"#, r#"%, World"#, '\\', true),
-            (r#"test"#, r#"te%st"#, '\\', true),
-            (r#"test"#, r#"te%%st"#, '\\', true),
-            (r#"test"#, r#"test%"#, '\\', true),
-            (r#"test"#, r#"%test%"#, '\\', true),
-            (r#"test"#, r#"t%e%s%t"#, '\\', true),
-            (r#"test"#, r#"_%_%_%_"#, '\\', true),
-            (r#"test"#, r#"_%_%st"#, '\\', true),
-            (r#"C:"#, r#"%\"#, '\\', false),
-            (r#"C:\"#, r#"%\"#, '\\', true),
-            (r#"C:\Programs"#, r#"%\"#, '\\', false),
-            (r#"C:\Programs\"#, r#"%\"#, '\\', true),
-            (r#"C:"#, r#"%\\"#, '\\', false),
-            (r#"C:\"#, r#"%\\"#, '\\', true),
-            (r#"C:\Programs"#, r#"%\\"#, '\\', false),
-            (r#"C:\Programs\"#, r#"%\\"#, '\\', true),
-            (r#"C:\Programs\"#, r#"%Prog%"#, '\\', true),
-            (r#"C:\Programs\"#, r#"%Pr_g%"#, '\\', true),
-            (r#"C:\Programs\"#, r#"%%\"#, '%', true),
-            (r#"C:\Programs%"#, r#"%%%"#, '%', true),
-            (r#"C:\Programs%"#, r#"%%%%"#, '%', true),
-            (r#"hello"#, r#"\%"#, '\\', false),
-            (r#"%"#, r#"\%"#, '\\', true),
-            (r#"3hello"#, r#"%%hello"#, '%', true),
-            (r#"3hello"#, r#"3%hello"#, '3', false),
-            (r#"3hello"#, r#"__hello"#, '_', false),
-            (r#"3hello"#, r#"%_hello"#, '%', true),
-        ];
-        let mut ctx = EvalContext::default();
-        for (target_str, pattern_str, escape, exp) in cases {
-            let target = datum_expr(Datum::Bytes(target_str.as_bytes().to_vec()));
-            let pattern = datum_expr(Datum::Bytes(pattern_str.as_bytes().to_vec()));
-            let escape = datum_expr(Datum::I64(escape as i64));
-            let op = scalar_func_expr(ScalarFuncSig::LikeSig, &[target, pattern, escape]);
-            let op = Expression::build(&mut ctx, op).unwrap();
-            let got = op.eval(&mut ctx, &[]).unwrap();
-            let exp = Datum::from(exp);
-            assert_eq!(got, exp, "{:?} like {:?}", target_str, pattern_str);
         }
     }
 }
