+++ src/coprocessor/dag/expr/mod.rs
@@ -16,6 +16,7 @@ mod builtin_cast;
 mod builtin_compare;
 mod builtin_control;
 mod builtin_json;
+mod builtin_like;
 mod builtin_math;
 mod builtin_op;
 mod builtin_time;
